Ignore snapshot destruction errors
The combination of `zfs send -R` and `zfs recv -F` deletes snapshots on
the destination that do not exist on the source.  In some scenarios,
these are the only snapshots that need to be deleted.  When that
happens, the `zfs destroy` fails because none of the snapshots exist.
We now ignore that error.