Eliminate shell=True
Instead of doing subprocess.Popen("zfs send ... | ssh zfs recv"),
we now create two separate subprocesses and connect them together.
This is safer, at least in theory, as shell=True can be dangerous.
Also, we now pass the arguments as a list rather than a space-delimited
string, as there is no shell to split the arguments.