Only use -R with -s
-s is a newer zfs send option that skips (rather than failing) on
missing snapshots on descendants.  For example, if you are sending
dataset tank/a@foo, which has a child tank/a/b, and tank/a/b@foo does
not exist, a plain -R will fail, but -R -s will not.

While requiring -s means we can no longer generate replication stream
packages on older systems, I don't think that's a huge downside.  First
of all, to be honest, I'm not sure that we care that much about
replication stream packages anyway; we already have to handle the case
where can't use them.  Second, it's better to work at all than fail.