Fix reading output from processes
We had a check for "cannot resume send", but it was having no effect.
This is because the send process's stderr was not connected to a PIPE.
As a result, the script would not notice this error message and thus
would not flag the resume as having failed, and thus would fail rather
than aborting the interrupted recv.

The fix is complicated because we need to read from two pipes (the
stderr of the send process and stdout of the recv process).  This means
we need to use threads and queues, as described here (and some Stack
Overflow posts):
https://lyceum-allotments.github.io/2017/03/python-and-pipes-part-6-multiple-subprocesses-and-pipes/

In that example, the threads were set to daemon mode, which let the
Python process exit when the main thread (which is non-daemon) was done.
Instead of that, I have the send and receive threads be non-daemon (the
default) and explicitly exit when the process they are monitoring is
done.  Then, the main thread only needs to watch the status of the child
threads.  This seems more correct.

Since we are doing the same thing with the output from both threads, we
can use a single queue, instead of the per-thread queue the example
uses.  This simplies the processing of the results in the main thread.

We also now check the exit status of both processes.